requirements.txt
pandas==2.2.2
numpy==1.26.4
python-telegram-bot==20.6
ccxt==4.3.0
pandas_ta==0.3.14b0